rename test-protobuf folder
